Thank you to the one who made it possible!
And no... it will never be those one folks claiming to have given me something!
Truth is... I earned what they 'gave'! So, that makes it not-a-gift!
When you keep a score card and count these 'gifts' so you can collect on them later...
That is not-a-gift! 
Friends don't do that to friends! 
mmmmk